MEMO AI
Memo Ai is an untrained model which is used to sort the task of a new beginner in using computer system. it helps to do the simple as well as some complex task which will be helpful to the user in operating computer system and also helps user to get the overview of the computer system or understand it.